+++ CHANGELOG.md
@@ -8,15 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ## [3.17.2] - 2019-02-18
+### Fixed
+- Fix product images carousel refresh bug.
 
 ## [3.17.1] - 2019-02-18
 ### Fixed
 - Quick fix on `AutocompleteInput` to remove warnings.
 ### Added
 - Base settings for tests.
-
-### Fixed
-- Fix product images carousel refresh bug.
 
 ## [3.17.0] - 2019-02-15
 ### Added