+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Snapshot tests for ProductPrice.
 - Snapshot tests for `BuyButton`
 
 ## [3.17.2] - 2019-02-18
@@ -26,7 +27,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Base settings for tests.
-- Snapshot tests for ProductPrice.
 
 ## [3.17.0] - 2019-02-15