+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Snapshot tests for `CollectionBadges`.
 - Snapshot tests for `Container`.
 - Snapshot tests for `DiscountBadge`.
+- Snapshot tests for `CategoriesHighlights.test`.
 
 ## [3.17.2] - 2019-02-18