+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Snapshot tests for `CategoriesHighlights`.
 - Snapshot tests for `GradientCollapse`.
 - Snapshot tests for `Greeting`.
+- Snapshot tests for `Logo`.
 
 ## [3.17.2] - 2019-02-18