+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Better use of helpers in tests.
 - Snapshot tests for `ProductDescription`.
 - Snapshot tests for `Logo`.
+- Snapshot tests for `ProductSpecifications`.
 
 ## [3.17.2] - 2019-02-18