+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.18.0] - 2019-02-21
+
 ### Added
 
 - Snapshot tests for `ProductPrice`.
@@ -27,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Snapshot tests for `SKUSelector`.
 - Snapshot tests for `ShippingSimulator`.
 - Snapshot tests for `SearchBar`.
+- Create `InfoCard` component
 
 ## [3.17.2] - 2019-02-18